refactor: moved extra config files
