--- config/components.json
@@ -1,5 +1,0 @@
-[
-  "@elifesciences/xpub-meca-export",
-  "@elifesciences/xpub-server",
-  "@pubsweet/component-send-email"
-]
