+++ config/staging.js
@@ -1,4 +1,4 @@
-const logger = require('./logger')
+const logger = require('./non-serializable/logger')
 
 module.exports = {
   'pubsweet-server': {
